Seperate Questions from QuestionsHandlers, make a class for Question, add new functions for questions
Now all the questions are not one big Enum, but a dictionary of IDs and objects of the question class.

Each question now has fields that are responsible for the type of question, the availability of answers, the id of the next and previous question, as well as the id of the parent question, if any. 

Added functions for getting the current, previous, next question by id, getting the type of question. 

The types of questions are separated into a separate Enum.